requirements.txt
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
scikit_learn==1.2.2
streamlit==1.20.0
pickle-mixin==1.0.2
